+++ cmn-data-svc-client/src/main/java/org/acumos/cds/client/CommonDataServiceRestClientMockImpl.java
@@ -62,7 +62,6 @@ import org.acumos.cds.domain.MLPSolutionFavorite;
 import org.acumos.cds.domain.MLPSolutionGroup;
 import org.acumos.cds.domain.MLPSolutionRating;
 import org.acumos.cds.domain.MLPSolutionRevision;
-import org.acumos.cds.domain.MLPSolutionWeb;
 import org.acumos.cds.domain.MLPStepResult;
 import org.acumos.cds.domain.MLPStepStatus;
 import org.acumos.cds.domain.MLPStepType;
@@ -162,7 +161,6 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 	private MLPUserNotifPref usrNotifPref;
 	private List<MLPUserNotifPref> userNotifPreferences;
 	private MLPStepResult stepResult;
-	private MLPSolutionWeb solutionWeb = new MLPSolutionWeb();
 	private List<MLPUser> solutionAccessUsers;
 	private RestPageResponse<MLPSolution> userAccessSolutions;
 	private RestPageResponse<MLPSolutionDeployment> userDeployments;
@@ -1171,15 +1169,6 @@ public class CommonDataServiceRestClientMockImpl implements ICommonDataServiceRe
 	@Override
 	public void updateUserNotificationPreference(MLPUserNotifPref usrNotifpref) {
 		// How to mock?
-	}
-
-	public void setSolutionWebMetadata(MLPSolutionWeb web) {
-		this.solutionWeb = web;
-	}
-
-	@Override
-	public MLPSolutionWeb getSolutionWebMetadata(String solutionId) {
-		return solutionWeb;
 	}
 
 	public void setSolutionAccessUsers(List<MLPUser> users) {