+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/ClientMethodTest.java
@@ -726,11 +726,6 @@ public class ClientMethodTest {
 			logger.info("Client failed as expected: {}", ex.toString());
 		}
 		try {
-			client.getSolutionWebMetadata("solutionId");
-		} catch (ResourceAccessException ex) {
-			logger.info("Client failed as expected: {}", ex.toString());
-		}
-		try {
 			client.getSolutionAccessUsers("solutionId");
 		} catch (ResourceAccessException ex) {
 			logger.info("Client failed as expected: {}", ex.toString());