+++ cmn-data-svc-client/src/test/java/org/acumos/cds/client/test/MockClientTest.java
@@ -48,7 +48,6 @@ import org.acumos.cds.domain.MLPSolutionFavorite;
 import org.acumos.cds.domain.MLPSolutionGroup;
 import org.acumos.cds.domain.MLPSolutionRating;
 import org.acumos.cds.domain.MLPSolutionRevision;
-import org.acumos.cds.domain.MLPSolutionWeb;
 import org.acumos.cds.domain.MLPStepResult;
 import org.acumos.cds.domain.MLPTag;
 import org.acumos.cds.domain.MLPThread;
@@ -357,10 +356,6 @@ public class MockClientTest {
 		client.addUserToNotification("id", "id");
 		client.dropUserFromNotification("id", "id");
 		client.setUserViewedNotification("id", "id");
-
-		MLPSolutionWeb web = new MLPSolutionWeb();
-		client.setSolutionWebMetadata(web);
-		Assert.assertTrue(web == client.getSolutionWebMetadata("id"));
 
 		List<MLPUser> userList = new ArrayList<>();
 		client.setSolutionAccessUsers(userList);